client: move plugins to libs
